delete is done

+++ public/main.js
@@ -116,7 +116,8 @@ function getCanvasImg(){
         pic.classList.add('picture');
         pic.setAttribute('data-id',element.id);
         console.log('the ID of',element.picName,'is:', element.id);
-        $('.profileSpace').prepend(pic);
+        $('.profileSpace').append(pic);
+        $('.profileSpace').append('<button class="delThis" onclick="deleteThis(this)" data-id="' + element.id + '">delete</button>')
       });
     }
   });
@@ -133,26 +134,16 @@ function deleteImg(id){
   });
 }
 
-$('#deletePage').on('click', '.picture', function(){
-  console.log('you can now delete');
-  deleleThis();
-});
+function deleteThis(elem){
+  if(confirm('are you sure you want to delete this image?')){  var id = $(elem).data('id');
+    console.log('this is the id',id);
+    $("*[data-id="+id+"]").remove();
+    deleteImg(id);
+  }else{
+    alert('image not deleted');
+  }
 
-$('#deletePace').off('click');
-
-function deleteThis(){
-  $('.profileSpace').on('click', '.picture', function(event){
-      event.preventDefault();
-      var id = $(this).data('id');
-      $(this).remove();
-      deleteImg(id);
-     console.log(id);
-  });
 }
-
-
-
-
 
 $('#logIn').submit(function(event){
   event.preventDefault();
